project_work

+++ acquire.py
@@ -10,11 +10,12 @@ def get_zillow_data():
         return pd.read_csv(filename)
     
     else:
-        sql = '''
+        sql = """
         SELECT parcelid, bedroomcnt, bathroomcnt, calculatedfinishedsquarefeet, taxvaluedollarcnt, yearbuilt, taxamount, fips, assessmentyear, landtaxvaluedollarcnt, lotsizesquarefeet, latitude, longitude
         FROM properties_2017
-        WHERE propertylandusetypeid = 261;
-        '''
+        JOIN predictions_2017 USING (parcelid)
+        WHERE transactiondate LIKE '2017%%'  and propertylandusetypeid = 261;
+        """
 
         df = pd.read_sql(sql, get_db_url('zillow'))
 
@@ -29,17 +30,45 @@ def get_zillow_locs():
         return pd.read_csv(filename)
 
     else:
-        sql = '''
+        sql = """
         SELECT parcelid, latitude, longitude
         FROM properties_2017
         WHERE propertylandusetypeid = 261;
-        '''
+        """
+
+        df = pd.read_sql(sql, get_db_url('zillow'))
+
+        df.to_csv(filename)
+
+        return df
+
+def get_zillow_zips():
+    filename = 'train_zips.csv'
+
+    if os.path.isfile(filename):
+        return pd.read_csv(filename)
+
+def get_zillow_other():
+    filename = 'zillow_other.csv'
+    
+    if os.path.isfile(filename):
+        return pd.read_csv(filename)
+    
+    else:
+        sql = """
+        SELECT parcelid, fireplacecnt, garagecarcnt, poolcnt
+        FROM properties_2017
+        JOIN predictions_2017 USING (parcelid)
+        WHERE transactiondate LIKE '2017%%'  and propertylandusetypeid = 261;
+        """
 
         df = pd.read_sql(sql, get_db_url('zillow'))
 
         df.to_csv(filename)
 
         return df 
+
+
